// src/api/BeatSaver/Types.d.ts
// Auto Generated by https://jvilk.com/MakeTypes/
export interface BeatSaverMap {
	id: string;
	name: string;
	description: string;
	uploader: Uploader;
	metadata: Metadata;
	stats: Stats;
	uploaded: string;
	automapper: boolean;
	ranked: boolean;
	qualified: boolean;
	versions?: VersionsEntity[] | null;
}

export interface Uploader {
	id: number;
	name: string;
	hash: string;
	avatar: string;
}
export interface Metadata {
	bpm: number;
	duration: number;
	songName: string;
	songSubName: string;
	songAuthorName: string;
	levelAuthorName: string;
}
export interface Stats {
	plays: number;
	downloads: number;
	upvotes: number;
	downvotes: number;
	score: number;
}
export interface VersionsEntity {
	hash: string;
	key: string;
	state: string;
	createdAt: string;
	sageScore: number;
	diffs?: DiffsEntity[] | null;
	downloadURL: string;
	coverURL: string;
	previewURL: string;
}
export interface DiffsEntity {
	njs: number;
	offset: number;
	notes: number;
	bombs: number;
	obstacles: number;
	nps: number;
	length: number;
	characteristic: "Standard" | "OneSaber" | "NoArrows" | "90Degree" | "360Degree" | "Lightshow" | "Lawless";
	difficulty: "Easy" | "Normal" | "Hard" | "Expert" | "ExpertPlus";
	events: number;
	chroma: boolean;
	me: boolean;
	ne: boolean;
	cinema: boolean;
	seconds: number;
	paritySummary: ParitySummary;
}
export interface ParitySummary {
	errors: number;
	warns: number;
	resets: number;
}